feat: Add PayU Biz payment integration
The docs for PayU and ActiverMerchant offsite are outdated and/or give insecure
methods. Instead, send a POST request to the PayU url with the required params.
The callback urls 'payment/success' and 'payment/failure' are configured to skip
auth_token checking. No current way to redirect using payment option (CC/NB),
so will be removed in the future.

The success url and database updates are NOT TESTED.