feat: Add patient lookup and creation through API call
The API call utilizes an API key to send a post query the
doctor side API endpoint. The system maintains it's own database
for new patients and creates a new entry iff it doesn't already
exist in the doctor side. It then sends a create request to the
doctor side.

This keeps both applications flexible and reduces coupling.
This commit also adds mail support setup.